refactor: move mgmt hooks into dedicated file
