+++ src/services/mgmt/index.ts
@@ -1,3 +1,6 @@
 import useTrackingToken from "./useTrackingToken";
+import useMgmtDefinition from "./useMgmtDefinition";
+import useUser from "./useUser";
+import useUpdateUser from "./useUpdateUser";
 
-export { useTrackingToken };
+export { useTrackingToken, useMgmtDefinition, useUser, useUpdateUser };
