+++ src/services/mgmt/useMgmtDefinition.ts
@@ -1,0 +1,18 @@
+import { useQuery } from "react-query";
+import { parse } from "yaml";
+import { listRepoFileContent } from "@/lib/github";
+
+const useMgmtDefinition = () => {
+  return useQuery(["mgmt", "encoded-definition"], async (): Promise<string> => {
+    const { content } = await listRepoFileContent(
+      "instill-ai",
+      "mgmt-backend",
+      "protocol/mgmt_protocol.yaml"
+    );
+
+    const decode = window.atob(content);
+    return Promise.resolve(parse(decode));
+  });
+};
+
+export default useMgmtDefinition;
