+++ src/services/mgmt/useUpdateUser.ts
@@ -1,0 +1,19 @@
+import { useMutation, useQueryClient } from "react-query";
+import { updateLocalUserMutation, User } from "@/lib/instill";
+
+const useUpdateUser = () => {
+  const queryClient = useQueryClient();
+  return useMutation(
+    async (payload: Partial<User>) => {
+      const user = await updateLocalUserMutation(payload);
+      return Promise.resolve(user);
+    },
+    {
+      onSuccess: (newUser) => {
+        queryClient.setQueryData<User>(["user", newUser.id], newUser);
+      },
+    }
+  );
+};
+
+export default useUpdateUser;
